catalog_builder_flat_root/requirements.txt
pandas==2.2.2
jinja2==3.1.4

duckduckgo_search==6.2.6
requests==2.32.3
Pillow==10.4.0
